move testing utils to testing
