+++ quic/s2n-quic-transport/src/stream/mod.rs
@@ -30,4 +30,4 @@ pub type StreamManager = AbstractStreamManager<StreamImpl>;
 // Import all tests
 
 #[cfg(test)]
-mod tests;
+mod testing;
